current state pushed

+++ bot/src/main/java/xyz/dbotfactory/dbot/model/DebtReturnTransaction.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DebtReturnTransaction {
-    int fromId;
-    int toId;
+    long fromId;
+    long toId;
     double amount;
 }
